+++ src/App.vue
@@ -1,16 +1,7 @@
 <template>
-  <sluggo-navbar></sluggo-navbar>
   <router-view />
 </template>
 
-<script lang="ts">
-import SluggoNavbar from "@/components/SluggoNavbar.vue";
-
-export default {
-  components: {
-    SluggoNavbar
-  }
-};
-</script>
+<script lang="ts"></script>
 
 <style></style>
