+++ src/components/TeamWrapper.vue
@@ -1,12 +1,17 @@
 <template>
+  <sluggo-navbar/>
   <router-view />
 </template>
 
 <script lang="ts">
 import { defineComponent } from "vue";
+import SluggoNavbar from "@/components/navbar/SluggoNavbar.vue";
 
 const teamWrapperComponent = defineComponent({
-  name: "TeamWrapper"
+  name: "TeamWrapper",
+  components: {
+    SluggoNavbar
+  }
 });
 
 export default teamWrapperComponent;
