+++ src/components/navbar/index.ts
@@ -1,0 +1,3 @@
+import SluggoNavbar from "@/components/navbar/SluggoNavbar.vue";
+
+export default SluggoNavbar;
