+++ src/router/index.ts
@@ -6,7 +6,6 @@ import Admin from "@/views/Admin.vue";
 import Help from "@/views/Help.vue";
 import Auth from "@/views/Auth.vue";
 import Profile from "@/views/Profile.vue";
-// import TicketDetails from "@/views/TicketDetails.vue";
 import NewTeam from "@/views/NewTeam.vue";
 import TeamWrapper from "@/components/TeamWrapper.vue";
 
@@ -49,39 +48,8 @@ const routes: Array<RouteRecordRaw> = [
         name: "Home",
         component: Home
       },
-      // {
-      //   path: "tickets/:userId",
-      //   name: "TicketDetails",
-      //   component: TicketDetails
-      // }
     ]
   },
-  // {
-  //   path: "/",
-  //   name: "Home",
-  //   component: Home
-  // },
-  // {
-  //   path: "/tickets",
-  //   name: "Tickets",
-  //   component: Tickets
-  // },
-  // {
-  //   path: "/users",
-  //   name: "Users",
-  //   component: Users,
-  //   props: true
-  // },
-  // {
-  //   path: "/users/:id",
-  //   name: "Profile",
-  //   component: Profile
-  // },
-  // {
-  //   path: "/admin",
-  //   name: "Admin",
-  //   component: Admin
-  // },
   {
     path: "/help",
     name: "Help",
@@ -92,25 +60,11 @@ const routes: Array<RouteRecordRaw> = [
     name: "Auth",
     component: Auth
   },
-  // {
-  //   path: "/tickets/:id",
-  //   name: "TicketDetails",
-  //   component: TicketDetails
-  // },
   {
     path: "/new_team/",
     name: "NewTeam",
     component: NewTeam
   }
-  // {
-  //   path: "/about",
-  //   name: "About",
-  //   // route level code-splitting
-  //   // this generates a separate chunk (about.[hash].js) for this route
-  //   // which is lazy-loaded when the route is visited.
-  //   component: () =>
-  //     import(/* webpackChunkName: "about" */ "../views/About.vue")
-  // }
 ];
 
 const router = createRouter({
